Random formatting and variable increases to fix timeout 😶
Did you ever hear the tragedy of Darth Plagueis The Wise? No? I thought not. It's not a story the Jedi would tell you. It's a Sith legend. Darth Plagueis was a Dark Lord of the Sith, so powerful and so wise he could use the Force to influence the midichlorians to create life… He had such a knowledge of the dark side, he could even keep the ones he cared about from dying.